requirements.txt
numpy
Pillow
torch
torchvision
pyobjc-framework-Quartz
pytesseract
pynput
